Refactor download buttons to own component, use in analysis table
